Move utility android funcs to utility.go
